### 2025-02-25 #7845 @editorialbot
[PRE REVIEW]: pylevin: efficient numerical integration of integrals containing up to three Bessel functions
**Submitting author:** @Reischke (<a href="http://orcid.org/0000-0001-5404-8753">Robert Reischke</a>)
**Repository:** https://github.com/rreischke/levin_bessel
**Branch with paper.md** (empty if default branch): main
**Version:** v1.0.5
