+++ paper/paper.bib
@@ -12,15 +12,21 @@
 	pages = {949--970},
 }
 
-
-@software{karamanis_2021,
-       author = {{Karamanis}, Minas and {Beutler}, Florian},
-        title = "{hankl: Python implementation of the FFTLog algorithm for cosmology}",
- howpublished = {Astrophysics Source Code Library, record ascl:2112.022},
-         year = 2021,
-        month = dec,
-          eid = {ascl:2112.022},
-       adsurl = {https://ui.adsabs.harvard.edu/abs/2021ascl.soft12022K},
+@ARTICLE{leonard_2023,
+       author = {{Leonard}, C. Danielle and {Ferreira}, Tassia and {Fang}, Xiao and {Reischke}, Robert and {Schoeneberg}, Nils and {Tr{\"o}ster}, Tilman and {Alonso}, David and {Campagne}, Jean-Eric and {Lanusse}, Fran{\c{c}}ois and {Slosar}, An{\v{z}}e and {Ishak}, Mustapha},
+        title = "{The N5K Challenge: Non-Limber Integration for LSST Cosmology}",
+      journal = {The Open Journal of Astrophysics},
+     keywords = {Astrophysics - Cosmology and Nongalactic Astrophysics},
+         year = 2023,
+        month = feb,
+       volume = {6},
+          eid = {8},
+        pages = {8},
+          doi = {10.21105/astro.2212.04291},
+archivePrefix = {arXiv},
+       eprint = {2212.04291},
+ primaryClass = {astro-ph.CO},
+       adsurl = {https://ui.adsabs.harvard.edu/abs/2023OJAp....6E...8L},
       adsnote = {Provided by the SAO/NASA Astrophysics Data System}
 }
 
@@ -110,7 +116,7 @@ number = {1},
 pages = {95-101},
 year = {1996},
 issn = {0377-0427},
-doi = {https://doi.org/10.1016/0377-0427(94)00118-9},
+doi = {10.1016/0377-0427(94)00118-9},
 url = {https://www.sciencedirect.com/science/article/pii/0377042794001189},
 author = {David Levin},
 keywords = {Oscillatory integrals, Collocation},
