### 2025-02-25 #7845 @rreischke
Hi @danielskatz, that is correct, I am the author. I'm not sure what happened with the compilation because a PDF was generated for me. In any case, it works now.

Thanks a lot,
Robert
