### 2025-03-07 #7845 @jedbrown
Hi, @rreischke! At first glance, I'm not finding automated tests or community/contributing guidelines. Can you please check the [review checklist](https://joss.readthedocs.io/en/latest/review_checklist.html) and make sure you've hit all items before we recruit reviewers?
