### 2025-03-11 b326e9e
removing a lot of overhead, changing were the diagonal flag is set to the constructor and not the integrator call
