+++ src/pylevin.cpp
@@ -2,9 +2,11 @@
 #include <gsl/gsl_spline.h>
 #include <boost/math/special_functions/bessel.hpp>
 #include <boost/math/special_functions/chebyshev.hpp>
+#include <chrono>
 
-pylevin::pylevin(uint type_in, std::vector<double> x, const std::vector<std::vector<double>> &integrand, bool logx, bool logy, uint nthread)
+pylevin::pylevin(uint type_in, std::vector<double> x, const std::vector<std::vector<double>> &integrand, bool logx, bool logy, uint nthread, bool diagonal)
 {
+    is_diagonal = diagonal;
     if (integrand.size() != x.size())
     {
         throw std::range_error("support dimensions must match integrand dimensions");
@@ -46,15 +48,30 @@ pylevin::~pylevin()
     }
     if (system_of_equations_set)
     {
-#pragma omp parallel for num_threads(N_thread_max) schedule(auto)
-        for (uint i_integrand = 0; i_integrand < n_integrand; i_integrand++)
+        if (is_diagonal)
         {
-            for (uint i_variable = 0; i_variable < size_variables; i_variable++)
+#pragma omp parallel for num_threads(N_thread_max) schedule(auto)
+            for (uint i_integrand = 0; i_integrand < n_integrand; i_integrand++)
             {
-                for (uint i_bisec = 0; i_bisec < bisection[i_integrand][i_variable].size() - 1; i_bisec++)
+                for (uint i_bisec = 0; i_bisec < bisection[i_integrand].size() - 1; i_bisec++)
                 {
-                    gsl_matrix_free(LU_G_matrix[i_integrand][i_variable][i_bisec]);
-                    gsl_permutation_free(permutation[i_integrand][i_variable][i_bisec]);
+                    gsl_matrix_free(LU_G_matrix[i_integrand][i_bisec]);
+                    gsl_permutation_free(permutation[i_integrand][i_bisec]);
+                }
+            }
+        }
+        else
+        {
+#pragma omp parallel for num_threads(N_thread_max) schedule(auto)
+            for (uint i_integrand = 0; i_integrand < n_integrand; i_integrand++)
+            {
+                for (uint i_variable = 0; i_variable < size_variables; i_variable++)
+                {
+                    for (uint i_bisec = 0; i_bisec < bisection[i_integrand * size_variables + i_variable].size() - 1; i_bisec++)
+                    {
+                        gsl_matrix_free(LU_G_matrix[i_integrand * size_variables + i_variable][i_bisec]);
+                        gsl_permutation_free(permutation[i_integrand * size_variables + i_variable][i_bisec]);
+                    }
                 }
             }
         }
@@ -100,18 +117,9 @@ void pylevin::set_levin(uint n_col_in, uint maximum_number_bisections_in, double
     }
 }
 
-std::vector<std::vector<std::vector<double>>> pylevin::get_bisection()
-{
-    return bisection;
-}
-
 void pylevin::init_splines(std::vector<double> &x, const std::vector<std::vector<double>> &integrand, bool logx, bool logy)
 {
     n_integrand = integrand[0].size();
-    if (!bisection_set)
-    {
-        bisection.resize(n_integrand, std::vector<std::vector<double>>());
-    }
     if (!system_of_equations_set && !bisection_set)
     {
         spline_integrand.resize(integrand[0].size(), std::vector<gsl_spline *>(N_thread_max));
@@ -922,9 +930,18 @@ void pylevin::solve_LSE_single(double A, double B, uint col, uint i_integrand, d
 
     if (bisection_set && !system_of_equations_set)
     {
[…]
