### 2025-05-06 d50bdb3
adding absolute tolerance as possible input
