### 2025-06-12 #7845 @nkrusch
👋 @vissarion @hiepdang @zafeirakopoulos @yossibokorbleile would you be interested in reviewing this submission for JOSS?
