### 2025-06-16 #7845 @yossibokorbleile
Hey @nkrusch

Thanks for asking! Bessel functions are not something I'm very familiar with, and I'm quite busy before in go on vacation, so I can't do it before August. But if you still need someone to do it in August let me know.

Sorry,

Yossi
