### 2025-06-18 #7845 @nkrusch
Greetings, @heltonmc and @amimre,

would you be interested in reviewing this submission for JOSS?
