+++ README.md
@@ -40,6 +40,12 @@ and redo the ``pip`` installation.
 ### Testing the installation
 You can run ``pytest` to test the installation.
 
+
+## Video-Tutorial: Cosmology from Home 2025 recording
+Feel free to watch 
+[![my recording for Cosmology from Home 2025 on youtube](.docs/levin_talk.png)](https://www.youtube.com/watch?v=RcZVc6emnQU&t=2s)
+
+
 ## Tutorial: How to use the library
 There is a detailed documentation of the code on [levin-bessel.readthedocs.io](https://levin-bessel.readthedocs.io/en/latest/). For a tutorial, there is a detailed [python notebook](https://github.com/rreischke/levin_bessel/blob/main/tutorial/levin_tutorial.ipynb) in the ``tutorial`` directory
 
