### 2025-07-01 #7845 @nkrusch
@arthurmloureiro thanks for the soft accept.

The Journal of Open Source Software (JOSS) is a developer friendly journal for research software packages. The JOSS review is different from most other journals. We aim to work with the authors to help them meet our criteria instead of merely passing judgment on the submission. As such, the reviewers are encouraged to submit issues and pull requests on the software repository. As a reviewer, you will also have a checklist to guide you through the review steps. You will find the documentation for the review process here: https://joss.readthedocs.io/en/latest/reviewer_guidelines.html

Typically, a review period is 2-4 weeks. The review can start once I have confirmed at least two reviewers. So, at this time I will add you tentatively, and we wait for the second reviewer.
