### 2025-07-17 #7845 @nkrusch
Hello 👋  @andrewgiuliani @ma-sadeghi @KumarSaurabh1992,

I am looking for a reviewer for this submission. Would any of you be interested in reviewing for JOSS?
