### 2025-07-18 #8618 @editorialbot
[REVIEW]: pylevin: efficient numerical integration of integrals containing up to three Bessel functions
**Submitting author:** @rreischke (<a href="http://orcid.org/0000-0001-5404-8753">Robert Reischke</a>)
**Repository:** https://github.com/rreischke/levin_bessel
**Branch with paper.md** (empty if default branch): main
**Version:** v1.3a
