### 2025-07-23 #8618 @mithun218
@Reischke, Can I install it on Windows OS using Miniconda?
