### 2025-08-04 #7845 @danielskatz
> Hi [@nkrusch](https://github.com/nkrusch) thanks for your invitation. I am currently editing for JOSS and unfortunately I have no time for reviewing.

@vissarion - you might want to remove yourself from the JOSS reviewer database to avoid this in the future.
