### 2025-08-26 rreischke/levin_bessel#3 @mithun218
The Package does not work on Windows OS
I attempted to install the package on Windows but was unsuccessful. I believe it may be helpful to mention in the README that installation is only supported on macOS and other Linux-based systems.

### Related issues

openjournals/joss-reviews#8618
