### 2025-08-26 #8618 @mithun218
@rreischke, I attempted to install the package on Windows but was unsuccessful. I believe it may be helpful to mention in the README that installation is only supported on macOS and other Linux-based systems. Please see the following issue.
