### 2025-08-27 db45574
updated documentation and readme
