### 2025-08-27 rreischke/levin_bessel#3 @rreischke
Hi there, yes indeed, I only tested this for those two platforms. I added this caveat in the readme.
