### 2025-09-26 #8618 @nkrusch
Hello @rreischke,

just checking up on status. I have noted the two open issues. Beyond that, looks like we are mainly waiting for updates to the paper.  Do you have an estimate of when the paper changes might be ready?
