+++ paper/paper.md
@@ -86,8 +86,20 @@ It is therefore clear that `pylevin` is more accurate and around three to four o
 
 ![Speed and accuracy comparison of `pylevin` (shown in dashed red) against a standard adaptive quadrature (shown in solid blue). The runtime for the two methods is given in the legend. For the adaptive quadrature the maximum number of sub-intervals was set to 1000 (default is 50). The grey shaded region indicates when the quadrature starts to fail. The bottom panel shows the relative difference between the two methods.   **Left**: Result of the integral $I_2$. **Right**: Result of the integral $I_3$.  \label{fig:figure1}](paper_plot_joss.pdf)
 
-Additionally to this benchmark, `pylevin` was tested against some specialised methods estimating integrals including a single Bessel function such as Ogata's method @ogata_2005 with an implementation described in @murray_2019 or FFTLog-based methods [@hamilton_2000;@karamanis_2021;@leonard_2023] and found excellent agreement in the results and runtimes within a factor of two of the other methods.
 
+#Comparison with various codes
+Additionally to the benchmark, `pylevin` is compared with more specialised, codes who mostly solve integrals over single Bessel functions. All computing times presented here are averages over several runs. The comparison was done with an M3 processor with 8 cores (4 performance cores).
+
+We calculate the following Hankel transformation with `pylevin`and Ogata's method @ogata_2005 as implemented in the `hankel` package @murray_2019.
+
+$$
+\mathrm{integral}(k) = \int_0^\infty\frac{x^2}{x^2+1}J_0(kx)\;\mathrm{d}x\;,
+$$
+
+for 500 values of $k$ logarithmically-spaced between 1 and $10^4$. The result is depicted on the left side of \autoref{fig:figure2}. It can be seen that both methods agree very well and are roughly equally fast. While the Hankel transformation formally goes from zero to infinity, $a=10^{-5}$ and $b = 10^8$ was used for `pylevin`. This choice of course depends on the specific integrand. 
+
+
+![Comparison of `pylevin` with two methods to calculate a Hankel transformation. Dashed red is `pylevin` while solid blue is the alternative method. **Left**: Integral evaluated with the Ogata method using the `hankel` package. **Right**: Integral for the galaxy power spectrum monopole evaluated using the **hankl** package. Different lines refer to different redshifts.  \label{fig:figure2}](paper_plot_2_joss.pdf)
 
 
 # References
