+++ paper/paper.md
@@ -87,19 +87,31 @@ It is therefore clear that `pylevin` is more accurate and around three to four o
 ![Speed and accuracy comparison of `pylevin` (shown in dashed red) against a standard adaptive quadrature (shown in solid blue). The runtime for the two methods is given in the legend. For the adaptive quadrature the maximum number of sub-intervals was set to 1000 (default is 50). The grey shaded region indicates when the quadrature starts to fail. The bottom panel shows the relative difference between the two methods.   **Left**: Result of the integral $I_2$. **Right**: Result of the integral $I_3$.  \label{fig:figure1}](paper_plot_joss.pdf)
 
 
-#Comparison with various codes
+# Comparison with various codes
 Additionally to the benchmark, `pylevin` is compared with more specialised, codes who mostly solve integrals over single Bessel functions. All computing times presented here are averages over several runs. The comparison was done with an M3 processor with 8 cores (4 performance cores).
 
-We calculate the following Hankel transformation with `pylevin`and Ogata's method @ogata_2005 as implemented in the `hankel` package @murray_2019.
+## `hankel`
+We calculate the following Hankel transformation with `pylevin` and Ogata's method (@ogata_2005) as implemented in the `hankel` package (@murray_2019).
 
 $$
 \mathrm{integral}(k) = \int_0^\infty\frac{x^2}{x^2+1}J_0(kx)\;\mathrm{d}x\;,
+\label{fig:hankel}
 $$
 
 for 500 values of $k$ logarithmically-spaced between 1 and $10^4$. The result is depicted on the left side of \autoref{fig:figure2}. It can be seen that both methods agree very well and are roughly equally fast. While the Hankel transformation formally goes from zero to infinity, $a=10^{-5}$ and $b = 10^8$ was used for `pylevin`. This choice of course depends on the specific integrand. 
 
+## `hankl`
+Here, we follow the cosmology example provided in the `hankl` documentation (@karamanis_2021) to compute the monopole of the galaxy power spectrum:
 
-![Comparison of `pylevin` with two methods to calculate a Hankel transformation. Dashed red is `pylevin` while solid blue is the alternative method. **Left**: Integral evaluated with the Ogata method using the `hankel` package. **Right**: Integral for the galaxy power spectrum monopole evaluated using the **hankl** package. Different lines refer to different redshifts.  \label{fig:figure2}](paper_plot_2_joss.pdf)
+$$
+\xi_0(s) =\int_0^\infty\left(b^2+fb/3 +f^2/5\right)P_\mathrm{lin}(k)J_0(ks) k^2\;\mathrm{d}k\;,
+\label{eq:hankl}
+$$
+
+where $b$ is the galaxy bias, $f$ the logarithmic growth rate and $P_\mathrm{lin}(k)$ is the linear matter power spectrum, which is calculated using `camb` [@lewis_efficient_2000;@lewis_cosmological_2002;@howlett_cmb_2012] at six redshifts. Since `hankl` is FFT based, it requires $k$ to be discretised, the FFT-dual will then be calculated at the inverse grid points. For this comparison, we use $2^{10}$ logarithmically-spaced points between $k=10^{-4}$ and $k = 1$ for the transformation to converge. For `pylevin`, the number of points where the transformation is evaluated is arbitrary. Here we use 100 points, which is more than enough to resolve all features in $\xi_0$.
+The results are shown on the right of \autoref{fig:figure2} and good agreement can be found between the two methods with `hankl` being roughly twice as fast as `pylevin`.
+
[…]
