### 2025-10-08 153d828
paper update
