+++ paper/paper.md
@@ -100,16 +100,7 @@ $$
 
 for 500 values of $k$ logarithmically-spaced between 1 and $10^4$. The result is depicted on the left side of \autoref{fig:figure2}. It can be seen that both methods agree very well and are roughly equally fast. While the Hankel transformation formally goes from zero to infinity, $a=10^{-5}$ and $b = 10^8$ was used for `pylevin`. This choice of course depends on the specific integrand. 
 
-## `hankl`
-Here, we follow the cosmology example provided in the `hankl` documentation (@karamanis_2021) to compute the monopole of the galaxy power spectrum:
 
-$$
-\xi_0(s) =\int_0^\infty\left(b^2+fb/3 +f^2/5\right)P_\mathrm{lin}(k)J_0(ks) k^2\;\mathrm{d}k\;,
-\label{eq:hankl}
-$$
-
-where $b$ is the galaxy bias, $f$ the logarithmic growth rate and $P_\mathrm{lin}(k)$ is the linear matter power spectrum, which is calculated using `camb` [@lewis_efficient_2000;@lewis_cosmological_2002;@howlett_cmb_2012] at six redshifts. Since `hankl` is FFT based, it requires $k$ to be discretised, the FFT-dual will then be calculated at the inverse grid points. For this comparison, we use $2^{10}$ logarithmically-spaced points between $k=10^{-4}$ and $k = 1$ for the transformation to converge. For `pylevin`, the number of points where the transformation is evaluated is arbitrary. Here we use 100 points, which is more than enough to resolve all features in $\xi_0$.
-The results are shown on the right of \autoref{fig:figure2} and good agreement can be found between the two methods with `hankl` being roughly twice as fast as `pylevin`.
 
 ![Comparison of `pylevin` with two methods to calculate a Hankel transformation. Dashed red is `pylevin` while solid blue is the alternative method. **Left**: Integral in \autoref{eq:hankel} evaluated with the Ogata method using the `hankel` package. **Right**: Integral in \autoref{eq:hankl} for the galaxy power spectrum monopole evaluated using the **hankl** package. Different lines refer to different redshifts.  \label{fig:figure2}](paper_plot_2_joss.pdf)
 
