+++ paper/paper.bib
@@ -108,6 +108,17 @@ archivePrefix = {arXiv},
       adsnote = {Provided by the SAO/NASA Astrophysics Data System}
 }
 
+@article{lewis_cosmological_2002,
+	title = {Cosmological parameters from {CMB} and other data: {A} {Monte} {Carlo} approach},
+	volume = {D66},
+	doi = {10.1103/PhysRevD.66.103511},
+	journal = {Physical Review D},
+	author = {Lewis, Antony and Bridle, Sarah},
+	year = {2002},
+	pages = {103511},
+}
+
+
 @article{levin_1996,
 title = {Fast integration of rapidly oscillatory functions},
 journal = {Journal of Computational and Applied Mathematics},
