+++ paper/paper.bib
@@ -12,6 +12,26 @@
 	pages = {949--970},
 }
 
+@article{chisari_core_2019,
+	title = {Core {Cosmology} {Library}: {Precision} {Cosmological} {Predictions} for {LSST}},
+	volume = {242},
+	issn = {0067-0049},
+	shorttitle = {Core {Cosmology} {Library}},
+	url = {https://ui.adsabs.harvard.edu/abs/2019ApJS..242....2C},
+	doi = {10.3847/1538-4365/ab1658},
+	abstract = {The Core Cosmology Library (CCL) provides routines to compute basic cosmological observables to a high degree of accuracy, which have been verified with an extensive suite of validation tests. Predictions are provided for many cosmological quantities, including distances, angular power spectra, correlation functions, halo bias, and the halo mass function through state-of-the-art modeling prescriptions available in the literature. Fiducial specifications for the expected galaxy distributions for the Large Synoptic Survey Telescope (LSST) are also included, together with the capability of computing redshift distributions for a user-defined photometric redshift model. A rigorous validation procedure, based on comparisons between CCL and independent software packages, allows us to establish a well-defined numerical accuracy for each predicted quantity. As a result, predictions for correlation functions of galaxy clustering, galaxy-galaxy lensing, and cosmic shear are demonstrated to be within a fraction of the expected statistical uncertainty of the observables for the models and in the range of scales of interest to LSST. CCL is an open source software package written in C, with a Python interface and publicly available at https://github.com/LSSTDESC/CCL.},
+	urldate = {2024-02-21},
+	journal = {The Astrophysical Journal Supplement Series},
+	author = {Chisari, Nora Elisa and Alonso, David and Krause, Elisabeth and Leonard, C. Danielle and Bull, Philip and Neveu, Jérémy and Villarreal, Antonia Sierra and Singh, Sukhdeep and McClintock, Thomas and Ellison, John and Du, Zilong and Zuntz, Joe and Mead, Alexander and Joudaki, Shahab and Lorenz, Christiane S. and Tröster, Tilman and Sanchez, Javier and Lanusse, Francois and Ishak, Mustapha and Hlozek, Renée and Blazek, Jonathan and Campagne, Jean-Eric and Almoubayyed, Husni and Eifler, Tim and Kirby, Matthew and Kirkby, David and Plaszczynski, Stéphane and Slosar, Anže and Vrastil, Michal and Wagoner, Erika L. and {LSST Dark Energy Science Collaboration}},
+	month = may,
+	year = {2019},
+	note = {ADS Bibcode: 2019ApJS..242....2C},
+	keywords = {Astrophysics - Cosmology and Nongalactic Astrophysics, cosmology: theory, Astrophysics - Instrumentation and Methods for Astrophysics, large-scale structure of universe, dark energy},
+	pages = {2},
+	file = {Full Text PDF:/Users/robert/Zotero/storage/96Z6Y4GL/Chisari et al. - 2019 - Core Cosmology Library Precision Cosmological Pre.pdf:application/pdf},
+}
+
+
 @ARTICLE{leonard_2023,
        author = {{Leonard}, C. Danielle and {Ferreira}, Tassia and {Fang}, Xiao and {Reischke}, Robert and {Schoeneberg}, Nils and {Tr{\"o}ster}, Tilman and {Alonso}, David and {Campagne}, Jean-Eric and {Lanusse}, Fran{\c{c}}ois and {Slosar}, An{\v{z}}e and {Ishak}, Mustapha},
         title = "{The N5K Challenge: Non-Limber Integration for LSST Cosmology}",
