### 2025-10-28 #8618 @nkrusch
Some feedback on the paper:

* L40: missing space between the two words: `f_of_x`is
* Section Examples: the two code blocks use whitespace around symbols inconsistently, e.g.,
   * around `+` in `x**3 + (x**2 +x)`
   * around `,` in `(len(k), 1)` and `1e-3,1e4`
* L67: the LaTeX math block is not rendering correctly in `$ a = 10^ {-5}$`  => try `$a` (remove the excess space)
* L83: similar spacing issue as above with `$ k$`
