### 2025-11-05 476571e
paper corrections
