+++ paper/paper.md
@@ -42,7 +42,7 @@ The way `pylevin` works is that one first defines an integrand, $f(x)$, the inte
 
 ```python
 x = np.geomspace(1e-5,100,100)
-f_of_x = x**3 + (x**2 +x)
+f_of_x = x**3 + (x**2 + x)
 integral_type = 0 
 number_omp_threads = 1 
 interploate_logx = True
@@ -55,7 +55,7 @@ lp_single = levin.pylevin(integral_type,
                           number_omp_threads)
 ```
 
-Note that the broadcasting of `f_of_x`is required as one can, in principle, pass many different integrands at the same time, and the code always expects this dimension. We can then define the values $k$ and $\ell$ at which we want to evaluate the integral, which are all one-dimensional arrays of the same shape. Additionally, we also have to allocate the memory for the result, which is stored in-place:
+Note that the broadcasting of `f_of_x` is required as one can, in principle, pass many different integrands at the same time, and the code always expects this dimension. We can then define the values $k$ and $\ell$ at which we want to evaluate the integral, which are all one-dimensional arrays of the same shape. Additionally, we also have to allocate the memory for the result, which is stored in-place:
 
 ```python
 k = np.geomspace(1e-3,1e4,1000)
@@ -75,10 +75,10 @@ If we had passed more integrands before, the results must have the corresponding
 We use the following two integrals as an example:
 
 $$
-I_2 = \;\int_{10^{-5}}^{100} \mathrm{d}x \;(x^3 +x^2 +x)j_{10}(kx)j_5(kx)\;, 
+I_2 = \;\int_{10^{-5}}^{100} \mathrm{d}x \;(x^3 + x^2 + x)j_{10}(kx)j_5(kx)\;, 
 $$
 $$
-I_3 = \;\int_{10^{-5}}^{100} \mathrm{d}x \;(x^3 +x^2 +x)j_{10}(kx)j_5(kx)j_{15}(kx)\;,
+I_3 = \;\int_{10^{-5}}^{100} \mathrm{d}x \;(x^3 + x^2 + x)j_{10}(kx)j_5(kx)j_{15}(kx)\;,
 $$
 
 The result of $I_2$ is shown on the left and for $I_3$ on the right in \autoref{fig:figure1}. In order for the quadrature to converge over an extended $k$-range, the number of maximum sub-intervals was increased to $10^3$ ($2\times 10^3$) for $I_2$ ($I_3$). The grey-shaded area indicates where the quadrature fails to reach convergence even after this change. 
@@ -97,7 +97,7 @@ $$
 \mathrm{integral}(k) = \int_0^\infty\frac{x^2}{x^2+1}J_0(kx)\;\mathrm{d}x\;,
 $$
 
-for 500 values of $k$ logarithmically-spaced between 1 and $10^4$. The result is depicted on the left side of \autoref{fig:figure2}. It can be seen that both methods agree very well and are roughly equally fast. While the Hankel transformation formally extends from 0 to infinity, $ a = 10^ {-5}$ and $b = 10^8$ were used for `pylevin`. This choice of course depends on the specific integrand. 
+for 500 values of $k$ logarithmically-spaced between 1 and $10^4$. The result is depicted on the left side of \autoref{fig:figure2}. It can be seen that both methods agree very well and are roughly equally fast. While the Hankel transformation formally extends from 0 to infinity, $ a = 10^{-5}$ and $b = 10^8$ were used for `pylevin`. This choice of course depends on the specific integrand. 
 
 ## `hankl`
[…]
