+++ paper/paper.bib
@@ -173,24 +173,6 @@ archivePrefix = {arXiv},
 
 
 
-@ARTICLE{leonard_2023,
-       author = {{Leonard}, C. Danielle and {Ferreira}, Tassia and {Fang}, Xiao and {Reischke}, Robert and {Schoeneberg}, Nils and {Tr{\"o}ster}, Tilman and {Alonso}, David and {Campagne}, Jean-Eric and {Lanusse}, Fran{\c{c}}ois and {Slosar}, An{\v{z}}e and {Ishak}, Mustapha},
-        title = "{The N5K Challenge: Non-Limber Integration for LSST Cosmology}",
-      journal = {The Open Journal of Astrophysics},
-     keywords = {Astrophysics - Cosmology and Nongalactic Astrophysics},
-         year = 2023,
-        month = feb,
-       volume = {6},
-          eid = {8},
-        pages = {8},
-          doi = {10.21105/astro.2212.04291},
-archivePrefix = {arXiv},
-       eprint = {2212.04291},
- primaryClass = {astro-ph.CO},
-       adsurl = {https://ui.adsabs.harvard.edu/abs/2023OJAp....6E...8L},
-      adsnote = {Provided by the SAO/NASA Astrophysics Data System}
-}
-
 @ARTICLE{zieser_2016,
        author = {{Zieser}, Britta and {Merkel}, Philipp M.},
         title = "{The cross-correlation between 3D cosmic shear and the integrated Sachs-Wolfe effect}",
