+++ paper/paper.bib
@@ -1,6 +1,6 @@
 
 @article{ogata_2005,
-	title = {A Numerical Integration Formula Based on the Bessel Functions, Publications of the Research Institute for Mathematical Sciences},
+	title = {A Numerical Integration Formula Based on the {B}essel Functions},
 	volume = {41},
 	doi = {10.2977/prims/1145474602},
 	abstract = {},
