### 2025-11-11 ca4422a
case fixes in bib
