+++ paper/paper.bib
@@ -34,7 +34,7 @@
 
 @ARTICLE{leonard_2023,
        author = {{Leonard}, C. Danielle and {Ferreira}, Tassia and {Fang}, Xiao and {Reischke}, Robert and {Schoeneberg}, Nils and {Tr{\"o}ster}, Tilman and {Alonso}, David and {Campagne}, Jean-Eric and {Lanusse}, Fran{\c{c}}ois and {Slosar}, An{\v{z}}e and {Ishak}, Mustapha},
-        title = "{The N5K Challenge: Non-Limber Integration for LSST Cosmology}",
+        title = {The {N5K} Challenge: Non-Limber Integration for {LSST} Cosmology},
       journal = {The Open Journal of Astrophysics},
      keywords = {Astrophysics - Cosmology and Nongalactic Astrophysics},
          year = 2023,
